Fixed recursion error and added search as home page after authentication
