+++ templates/base.html
@@ -36,7 +36,6 @@
                     </div>
                 </div>
             </div>
-<<<<<<< HEAD
             <div class="footer-copyright">
                 <div class="container">
                     Made by <a class="orange-text text-lighten-3" href="http://materializecss.com">Knowty</a>
