+++ victim/views.py
@@ -8,12 +8,22 @@ from processor.utils import push_record_to_sqs_queue
 from victim.forms import RefugeeCreateForm
 from victim.models import Victim
 
+from django.contrib.auth.decorators import login_required
 
-class RefugeeCreateView(FormView):
+class LoginRequiredMixin(object):
+    @classmethod
+    def as_view(cls, **initkwargs):
+        view = super(LoginRequiredMixin, cls).as_view(**initkwargs)
+        return login_required(view)
+
+# Create your views here.
+
+class RefugeeCreateView(LoginRequiredMixin, FormView):
     """
     This will handle creation of Refugees.
     """
     template_name = 'victim/create.html'
+    login_url = 'login/'
     form_class = RefugeeCreateForm
     success_url = '/victim/success'
 
@@ -26,11 +36,11 @@ class RefugeeCreateView(FormView):
 
 
 
-class RefugeeSearchView(View):
+class RefugeeSearchView(LoginRequiredMixin, View):
     """
     This will handle Refugee search.
     """
-
+    login_url = 'login/'
     def get(self, request):
 
         type = 'search'
